azure_monitor/db_sample/requirements.txt
psycopg2 >= 2.9.1
azure-monitor-opentelemetry-exporter ~= 1.0.0
opentelemetry-instrumentation-psycopg2 >= 0.24b0
